feat: 상세 점수 추가
S12P21A304-319

+++ backend/src/main/java/newstock/domain/news/dto/TopNewsDto.java
@@ -14,7 +14,7 @@ public class TopNewsDto {
 
     private String publishedDate;
 
-    private int score;
+    private float score;
 
     public static TopNewsDto of(News news) {
         return TopNewsDto.builder()
